Rename Autowriter.RazorPages to Autowriter.UI
The name Autowriter.RazorPages refers to a specific implementation
detail - the fact that the UI is built using the RazorPages technology.
This isn't the purpose of the abstraction, the purpose is to separate
the UI from the domain. Furthermore, we could change the technology that
the UI is built with in the future. Therefore, I think that
Autowriter.UI is a better name.