Get latest upload for UI test
Changing the core DB from a singleton to be transient required changing
the tests from using an in-memory DB to one on the file system. As a
result, past test data has started to build up, and I cannot rely on a
clean start for my tests. Because of this, I am checking the last
uploaded piece of material so that the date matches